requirements.txt
appdirs==1.4.3
astroid==2.4.0
attrs==19.3.0
Automat==20.2.0
autopep8==1.5.2
beautifulsoup4==4.9.0
blis==0.4.1
bs4==0.0.1
catalogue==1.0.0
certifi==2020.4.5.1
cffi==1.14.0
cfscrape==2.1.1
chardet==3.0.4
click==7.1.1
colorama==0.4.3
constantly==15.1.0
cryptography==2.9.2
cssselect==1.1.0
cymem==2.0.3
facebook-scraper==0.1.12
fake-useragent==0.1.11
feedfinder2==0.0.4
feedparser==5.2.1
Flask==1.1.2
gunicorn==20.0.4
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.6.0
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
jieba3k==0.35.1
Jinja2==2.11.2
joblib==0.14.1
lazy-object-proxy==1.4.3
lxml==4.5.0
MarkupSafe==1.1.1
mccabe==0.6.1
murmurhash==1.0.2
newspaper3k==0.2.8
nltk==3.4.5
nodejs==0.1.1
numpy==1.18.2
optional-django==0.1.0
pandas==1.0.3
parse==1.15.0
parsel==1.6.0
Pillow==7.1.1
plac==1.1.3
preshed==3.0.2
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
pyee==7.0.1
pygame==1.9.6
PyHamcrest==2.0.2
pylint==2.5.0
pyOpenSSL==19.1.0
pyppeteer==0.0.25
pyquery==1.4.1
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
queuelib==1.5.0
requests==2.23.0
requests-file==1.4.3
requests-futures==1.0.0
requests-html==0.10.0
schedule==0.6.0
scikit-learn==0.22.2.post1
scipy==1.4.1
Scrapy==2.1.0
selenium==3.141.0
service-identity==18.1.0
six==1.14.0
sklearn==0.0
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
stem==1.8.0
thinc==7.4.0
tinysegmenter==0.3
tldextract==2.2.2
toml==0.10.0
torrequest==0.1.0
tqdm==4.45.0
Twisted==20.3.0
typed-ast==1.4.1
urllib3==1.25.8
w3lib==1.21.0
wasabi==0.6.0
webencodings==0.5.1
websockets==8.1
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.1.0
